efl_ui_smart_container: introduce a experimental new widget
this widget tries to replace efl.ui.stack efl.ui.flip & efl.ui.pager
In general those widgets do the same thing, they get content. And
display them with some sort of animations. The new idea here is, that
the basic widget smart_container only handles the ownership etc. of the
content that gets added to this. Then there is a transition object. The
transition object gets notified over the new contents, and requests for
displaying particular contents. The transition then handles those
things.

The version that is there right now is half way compatible with
Efl.Ui.Pager, and complatible with stack when the stack transition is
used.